[Feat] organize class
클래스 정리
