+++ CityFarmer/Assets/Script/Info/Food.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3cd9317e2d8efe24cb11e7850816b088
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
