+++ CityFarmer/Assets/Script/Info/Item.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 24f55e44f58afcd43b7e3521cc28371f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
